find elm-stuff relative to file; err if not found
elm-oracle must be run from the project path, otherwise it blows up
trying to figure out the dependencies.

so, find elm-stuff and cd to its parent dir before running elm-oracle.

if elm-stuff isn't found, print an error message, which is a bit
disruptive but lets the user figure out why their autocomplete isn't
working. would be nice if this was a warning instead, but that's hard to
do while in insert mode, since "-- INSERT --" covers up any echoed
messages.