Only show elm format error if syntastic is disabled
At present, syntax errors cause elm-format to fail outputting a nasty
red error which is immediately replaced by a syntastic error (presumably
saying exactly the same thing) in the quickfix box.

To prevent this I have added the additional condition that elm-format
only shows errors if syntastic is disabled.

I am not sure if this is a sensible solution to the issue, there are
conceivably errors that can occur with elm-format that aren't related to
syntax (i.e. filesystem errors)