Iteratively search down the segments
Previously if we had 'My.Module.Type' we'd fail as we'd think that Type
was the name of a module on disk and fail to find it.

Now we try the full thing, and if that fails, try it as a import alias,
if that fails, strip off the last segment, and try as a module, then an
import alias, etc, etc. Until there is nothing left to strip off and
then we error if we haven't found anything.